pyproject.toml
[project]
name = "leave-management-mcp"
version = "0.1.0"
description = "Leave manager application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastparquet>=2024.11.0",
    "mcp[cli]>=1.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
]
